zx/requirement.txt
astroid==1.4.6
backports.shutil-get-terminal-size==1.0.0
cffi==1.6.0
colorama==0.3.7
cryptography==1.3.2
Cython==0.24
decorator==4.0.9
Django==1.9
django-anymail==0.3.1
django-reversion==1.9.3
djangorestframework==3.3.3
gevent==1.1.1
greenlet==0.4.10
httplib2==0.9.2
idna==2.1
ipython==4.2.0
ipython-genutils==0.1.0
lazy-object-proxy==1.2.2
mistune==0.7.1
pickleshare==0.7.2
Pillow==3.2.0
pyasn1==0.1.9
pycparser==2.14
Pygments==2.0.2
pyOpenSSL==16.0.0
qrcode==5.2.2
requests==2.10.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.1
urllib3==1.14
virtualenv==15.0.1
wget==3.2
wrapt==1.10.8
yapf==0.8.0
you-get==0.4.365
